Added class existence check

+++ cmd.py
@@ -11,6 +11,15 @@ config = eval(f.read())
 f.close()
 
 
+def classExists(string):
+    try:
+        if type(eval(string))=="class":
+            return True
+    except:
+        return False
+    return False
+
+
 class basicCmds:
     def cmdExec(line):
         line = ' '.join(sys.argv[1:])
@@ -63,10 +72,13 @@ execClass = eval(config["execClass"])
 line = ' '.join(sys.argv[1:])
 
 if line[0:2]=="?#":
-    config["execClass"] = line[2:]
-    f = open(CWD + "/cmdConfig", "w")
-    f.write(str(config))
-    f.close()
-    print("Updated execution class")
+    if classExists(line[2:]):
+        config["execClass"] = line[2:]
+        f = open(CWD + "/cmdConfig", "w")
+        f.write(str(config))
+        f.close()
+        print("Updated execution class")
+    else:
+        print("No such class!")
 else:
     execClass.cmdExec(line)
